fixed HostInfo::generate method

+++ cnm_core/src/cnm/machine/host_info.cpp
@@ -9,7 +9,7 @@ HostInfo HostInfo::generate(std::string name, uint8_t f, uint8_t s, uint8_t t,
                             uint8_t fo) {
   std::stringstream ss;
 
-  // something is defenitely wrong with  C++ and it's uint8_t
+  // something is definitely wrong with  C++ and it's uint8_t
   // without this ugly type conversion stringstream treats uint8_t as
   // characters, not numbers.
   ss << (int)f << "." << (int)s << "." << (int)t << "." << (int)fo;
@@ -22,7 +22,7 @@ HostInfo HostInfo::generate(std::string name, uint8_t f, uint8_t s, uint8_t t,
 HostInfo HostInfo::generate(std::string name) {
   std::random_device rd;
   std::default_random_engine re(rd());
-  std::uniform_int_distribution<uint8_t> d(0, 255);
+  std::uniform_int_distribution<int> d(0, 255);
   return generate(name, d(re), d(re), d(re), d(re));
 }
